@@ -32,7 +32,7 @@
   "require": {
     "php": "^7.2",
     "contao-community-alliance/dc-general": "^2.2@dev",
-    "contao-community-alliance/events-contao-bindings": "~4.0",
+    "contao-community-alliance/events-contao-bindings": "^4.9@dev",
     "contao-community-alliance/meta-palettes": "~2.0",
     "contao-community-alliance/translator": "~2.1",
     "contao-community-alliance/url-builder": "~1.3",
@@ -42,7 +42,8 @@
     "menatwork/contao-multicolumnwizard-bundle": "^3.4",
     "ocramius/proxy-manager": "^1.0 || ^2.1",
     "symfony/config": "4.4.*",
-    "symfony/dependency-injection": "^3.4.26 || ^4.2.7",
+    "symfony/dependency-injection": "4.4.*",
+    "symfony/event-dispatcher": "4.4.*",
     "symfony/filesystem": "4.4.*",
     "symfony/finder": "4.4.*",
     "symfony/proxy-manager-bridge": "4.4.*",
@@ -53,7 +54,7 @@
     "contao/manager-plugin": "^2.8",
     "contao/manager-bundle": "^4.9.0, <4.13.0",
     "doctrine/doctrine-bundle": "^1.8",
-    "phpcq/all-tasks": "^1.2"
+    "phpcq/all-tasks": "^1.3"
   },
   "autoload": {
     "psr-4": {
