@@ -11,6 +11,7 @@
  * This project is provided in good faith and hope to be usable by anyone.
  *
  * @package    MetaModels/core
+ * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @copyright  2012-2020 The MetaModels team.
@@ -202,7 +203,7 @@ class BackendNavigationListener
      */
     private function extractLanguageValue($values, $locale): string
     {
-        return ($values[$locale] ?? $values['']);
+        return html_entity_decode(($values[$locale] ?? $values['']));
     }
 
     /**
