@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2020 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,7 +13,8 @@
  * @package    MetaModels/core
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -82,9 +83,7 @@ class BasicDefinitionBuilder
         // If we have variants, override all modes to tree mode.
         if ($metaModel->hasVariants()) {
             $config->setMode(BasicDefinitionInterface::MODE_HIERARCHICAL);
-        }
-
-        if ('hierarchical' === $meta['rendermode']) {
+        } elseif ('hierarchical' === $meta['rendermode']) {
             // Hierarchical mode - Records are displayed as tree (see site structure).
             $config->setMode(BasicDefinitionInterface::MODE_HIERARCHICAL);
         } elseif ('parented' === $meta['rendermode']) {
