@@ -123,8 +123,6 @@ class GetSearchablePagesListener implements ServiceAnnotationInterface
     /**
      * Start point for the hook getSearchablePages.
      *
-     * @Hook("getSearchablePages")
-     *
      * @param array       $pages       List with all pages.
      *
      * @param int|null    $rootPage    ID of the root page.
@@ -139,6 +137,8 @@ class GetSearchablePagesListener implements ServiceAnnotationInterface
      *
      * @see \RebuildIndex::run()
      * @see \Automator::generateSitemap()
+     *
+     * @Hook("getSearchablePages")
      *
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
@@ -161,7 +161,6 @@ class GetSearchablePagesListener implements ServiceAnnotationInterface
 
         asort($this->foundPages);
 
-        // Return the new list.
         return $this->foundPages;
     }
 
@@ -444,6 +443,49 @@ class GetSearchablePagesListener implements ServiceAnnotationInterface
     }
 
     /**
+     * Start point for the hook getSearchablePages.
+     *
+     * @param array       $pages       List with all pages.
+     *
+     * @param int|null    $rootPage    ID of the root page.
+     *
+     * @param bool|null   $fromSiteMap True when called from sitemap generator, null otherwise.
+     *
+     * @param string|null $language    The current language.
+     *
+     * @return array
+     *
+     * @throws \Doctrine\DBAL\DBALException When an database error occur.
+     *
+     * @see \RebuildIndex::run()
+     * @see \Automator::generateSitemap()
+     *
+     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     */
+    public function addPages($pages, $rootPage = null, $fromSiteMap = false, $language = null)
+    {
+        // Save the pages.
+        $this->foundPages = $pages;
+        unset($pages);
+
+        // Run each entry in the published config array.
+        foreach ($this->getConfigs() as $config) {
+            if (!$config['published']) {
+                continue;
+            }
+            $this->getMetaModelsPages(
+                $config,
+                $rootPage,
+                $language
+            );
+        }
+
+        asort($this->foundPages);
+
+        return $this->foundPages;
+    }
+
+    /**
      * Get a MetaModels, a filter and a renderSetting. Get all items based on the filter and build the jumpTo urls.
      *
      * @param array       $config   ID of the MetaModels.
@@ -472,7 +514,7 @@ class GetSearchablePagesListener implements ServiceAnnotationInterface
         $availableLanguages = $this->getLanguage($language, $metaModels);
         $currentLanguage    = $GLOBALS['TL_LANGUAGE'];
 
-        $foundPages = [];
+        $foundPages = [$this->foundPages];
 
         foreach ($availableLanguages as $newLanguage) {
             // Change language.
