@@ -16,6 +16,7 @@
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
+ * @author     Ingolf Steinhardt <info@e-spin.de>
  * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
@@ -112,11 +113,11 @@ class InsertTagsListener implements ServiceAnnotationInterface
     /**
      * Evaluate an insert tag.
      *
-     * @Hook("replaceInsertTags")
-     *
      * @param string $insertTag The tag to evaluate.
      *
      * @return bool|string
+     *
+     * @Hook("replaceInsertTags")
      */
     public function __invoke($insertTag)
     {
@@ -312,9 +313,9 @@ class InsertTagsListener implements ServiceAnnotationInterface
      *
      * @param int    $intID   ID of content element or module.
      *
-     * @return boolean|string Return false when nothing was found or the count value.
+     * @return int Return the count value.
      */
-    protected function getCount($strType, $intID)
+    protected function getCount($strType, $intID): int
     {
         switch ($strType) {
             // From module, can be a MetaModel list or filter.
@@ -329,16 +330,15 @@ class InsertTagsListener implements ServiceAnnotationInterface
 
             // Unknown element type.
             default:
-                return false;
+                return 0;
         }
 
         // Check if we have data.
         if (null !== $objMetaModelResult) {
-            return ($this->getCountFor($objMetaModelResult->metamodel, $objMetaModelResult->metamodel_filtering)
-                    ?? false);
+            return $this->getCountFor($objMetaModelResult->metamodel, $objMetaModelResult->metamodel_filtering);
         }
 
-        return false;
+        return 0;
     }
 
     /**
@@ -406,13 +406,13 @@ class InsertTagsListener implements ServiceAnnotationInterface
      *
      * @param int $intFilterId    ID of the filter.
      *
-     * @return boolean|null Null for no data or integer for the count result.
+     * @return int The count result.
      */
-    protected function getCountFor($intMetaModelId, $intFilterId): ?bool
+    protected function getCountFor($intMetaModelId, $intFilterId): int
     {
         $metaModel = $this->loadMetaModel($intMetaModelId);
         if (null === $metaModel) {
-            return null;
+            return 0;
         }
 
         $objFilter = $metaModel->getEmptyFilter();
