@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2020 The MetaModels team.
+ * (c) 2012-2021 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -18,7 +18,7 @@
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2020 The MetaModels team.
+ * @copyright  2012-2021 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -282,7 +282,7 @@ $GLOBALS['TL_DCA']['tl_metamodel_attribute'] = [
             'inputType' => 'checkbox',
             'eval'      => [
                 'submitOnChange' => true,
-                'tl_class'       => 'cbx w50'
+                'tl_class'       => 'w50 cbx m12'
             ],
             'sql'       => "char(1) NOT NULL default ''"
         ],
@@ -290,7 +290,7 @@ $GLOBALS['TL_DCA']['tl_metamodel_attribute'] = [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_attribute']['isunique'],
             'inputType' => 'checkbox',
             'eval'      => [
-                'tl_class' => 'cbx w50'
+                'tl_class' => 'w50 cbx m12'
             ],
             'sql'       => "char(1) NOT NULL default ''"
         ],
