@@ -68,6 +68,13 @@ class FilterBuilderSql
     private $connection;
 
     /**
+     * The table alias to use.
+     *
+     * @var string
+     */
+    private $tableAlias;
+
+    /**
      * Create a new instance.
      *
      * @param string     $tableName  The table name.
@@ -75,12 +82,15 @@ class FilterBuilderSql
      * @param string     $combiner   The combiner (AND or OR).
      *
      * @param Connection $connection The database connection.
+     *
+     * @param string     $tableAlias The table alias prefix (defaults to 't.').
      */
-    public function __construct($tableName, $combiner, $connection)
+    public function __construct($tableName, $combiner, $connection, string $tableAlias = 't.')
     {
         $this->tableName  = $tableName;
         $this->combiner   = strtoupper($combiner);
         $this->connection = $this->sanitizeConnection($connection);
+        $this->tableAlias = $tableAlias;
     }
 
     /**
@@ -142,7 +152,12 @@ class FilterBuilderSql
     protected function getFilterForComparingOperator($operation)
     {
         $this->parameter[]  = $operation['value'];
-        $this->procedures[] = sprintf('(%s %s ?)', $operation['property'], $operation['operation']);
+        $this->procedures[] = sprintf(
+            '(%s%s %s ?)',
+            $this->tableAlias,
+            $operation['property'],
+            $operation['operation']
+        );
 
         return $this;
     }
@@ -158,7 +173,8 @@ class FilterBuilderSql
     {
         $this->parameter    = array_merge($this->parameter, array_values($operation['values']));
         $this->procedures[] = sprintf(
-            '(%s IN (%s))',
+            '(%s%s IN (%s))',
+            $this->tableAlias,
             $operation['property'],
             rtrim(str_repeat('?,', \count($operation['values'])), ',')
         );
@@ -178,7 +194,7 @@ class FilterBuilderSql
     protected function getFilterForLike($operation)
     {
         $this->parameter[]  = str_replace(array('*', '?'), array('%', '_'), $operation['value']);
-        $this->procedures[] = sprintf('(%s LIKE ?)', $operation['property']);
+        $this->procedures[] = sprintf('(%s%s LIKE ?)', $this->tableAlias, $operation['property']);
 
         return $this;
     }
