@@ -18,6 +18,7 @@
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
  * @author     Cliff Parnitzky <github@cliff-parnitzky.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
+ * @author     Ingolf Steinhardt <info@e-spin.de>
  * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
@@ -184,7 +185,7 @@ class Helper
         $extra = $property->getExtra();
 
         $extra['minCount']       =
-        $extra['maxCount'] = count($languages);
+        $extra['maxCount']       = count($languages);
         $extra['disableSorting'] = true;
         $extra['tl_class']       = 'clr w50';
         $extra['columnFields']   = [
