@@ -160,7 +160,7 @@ abstract class MetaModelHybrid extends Hybrid
             $strInfo = '';
             if ($this->metamodel) {
                 // Add CSS file.
-                $GLOBALS['TL_CSS'][] = 'system/modules/metamodels/assets/css/style.css';
+                $GLOBALS['TL_CSS'][] = 'bundles/metamodelscore/css/style.css';
 
                 // Retrieve name of MetaModels.
                 $infoTemplate =
