@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2021 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -14,12 +14,14 @@
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     David Maack <david.maack@arcor.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2021 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
 namespace MetaModels\Helper;
+
+use Contao\Controller;
 
 /**
  * Gateway to the Contao "Controller" class for usage of the core without
@@ -31,7 +33,7 @@ namespace MetaModels\Helper;
  *
  * @deprecated Deprecated in favor of contao-events-bindings
  */
-class ContaoController extends \Controller
+class ContaoController extends Controller
 {
     /**
      * The instance.
