@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2020 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -14,7 +14,8 @@
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @author     Ingolf Steinhardt <info@e-spin.de>
+ * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -195,6 +196,7 @@ class InputScreenInformationBuilder
                 // @codingStandardsIgnoreEnd
                 return $column;
             }
+
             $column = array_merge(
                 $column,
                 $attribute->getFieldDefinition($column),
@@ -308,7 +310,7 @@ class InputScreenInformationBuilder
                 $label[$availableLanguage] = $metaModel->getName();
             }
         } else {
-            $label[$metaModel->getActiveLanguage()] = $metaModel->getName();
+            $label['default'] = $metaModel->getName();
         }
 
         $legend = [
@@ -365,7 +367,7 @@ class InputScreenInformationBuilder
             $result['legend' . (\count($result) + 1)] = $legend;
         }
         if (null === $fallback) {
-            $label = ['' => $property['legendtitle']];
+            $label = ['default' => $property['legendtitle']];
         } else {
             $label            = unserialize($property['legendtitle'], ['allowed_classes' => false]);
             $label['default'] = $label[$fallback];
