@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2021 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -14,7 +14,7 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2021 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -31,6 +31,8 @@ use PHPUnit\Framework\TestCase;
 
 /**
  * Class BaseSimpleTest
+ *
+ * @covers \MetaModels\Attribute\BaseSimple
  */
 class BaseSimpleTest extends TestCase
 {
@@ -47,8 +49,8 @@ class BaseSimpleTest extends TestCase
 
         $attribute = new BaseSimple($metaModel, [], $connection, $manipulator);
 
-        $this->assertInstanceOf(BaseSimple::class, $attribute);
-        $this->assertInstanceOf(IAttribute::class, $attribute);
-        $this->assertInstanceOf(ISimple::class, $attribute);
+        self::assertInstanceOf(BaseSimple::class, $attribute);
+        self::assertInstanceOf(IAttribute::class, $attribute);
+        self::assertInstanceOf(ISimple::class, $attribute);
     }
 }
