@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2021 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,7 +13,7 @@
  * @package    MetaModels/core
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2021 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -44,19 +44,19 @@ class BaseTest extends TestCase
         $metaModel = $this->getMockForAbstractClass(IMetaModel::class);
 
         $metaModel
-            ->expects($this->any())
+            ->expects(self::any())
             ->method('getTableName')
-            ->will($this->returnValue('mm_unittest'));
+            ->willReturn('mm_unittest');
 
         $metaModel
-            ->expects($this->any())
+            ->expects(self::any())
             ->method('getActiveLanguage')
-            ->will($this->returnValue($language));
+            ->willReturn($language);
 
         $metaModel
-            ->expects($this->any())
+            ->expects(self::any())
             ->method('getFallbackLanguage')
-            ->will($this->returnValue($fallbackLanguage));
+            ->willReturn($fallbackLanguage);
 
         return $metaModel;
     }
@@ -146,10 +146,10 @@ class BaseTest extends TestCase
         $attribute = $this->getAttribute($attributes);
 
         foreach ($attributes as $key => $value) {
-            $this->assertEquals($value, $attribute->get($key), $key);
+            self::assertEquals($value, $attribute->get($key), $key);
         }
 
-        $this->assertEquals('baseattribute', $attribute->getColName());
+        self::assertEquals('baseattribute', $attribute->getColName());
     }
 
     /**
@@ -162,7 +162,7 @@ class BaseTest extends TestCase
         /** @var Base $attribute */
         $attribute = $this->getAttribute(array('foo' => 'bar'));
 
-        $this->assertEquals(null, $attribute->get('foo'));
+        self::assertEquals(null, $attribute->get('foo'));
     }
 
     /**
@@ -182,10 +182,10 @@ class BaseTest extends TestCase
             )
         );
 
-        $this->assertFalse(array_key_exists('filter', $fieldDefinition['eval']));
-        $this->assertFalse(array_key_exists('search', $fieldDefinition['eval']));
+        self::assertFalse(array_key_exists('filter', $fieldDefinition['eval']));
+        self::assertFalse(array_key_exists('search', $fieldDefinition['eval']));
 
-        $this->assertEquals('some_widget_class', $fieldDefinition['eval']['tl_class']);
-        $this->assertEquals(true, $fieldDefinition['eval']['readonly']);
+        self::assertEquals('some_widget_class', $fieldDefinition['eval']['tl_class']);
+        self::assertEquals(true, $fieldDefinition['eval']['readonly']);
     }
 }
