@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2021 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -13,7 +13,7 @@
  * @package    MetaModels/core
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2021 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -87,7 +87,7 @@ class ConditionAndTest extends TestCase
             $condition->addChild($this->mockFilter($filter));
         }
 
-        $this->assertEquals($expected, $condition->getMatchingIds(), $message);
+        self::assertEquals($expected, $condition->getMatchingIds(), $message);
     }
 
     /**
@@ -99,7 +99,7 @@ class ConditionAndTest extends TestCase
      */
     private function mockFilter($result)
     {
-        $filter = $this->getMockForAbstractClass('MetaModels\\Filter\\IFilter');
+        $filter = $this->getMockForAbstractClass(IFilter::class);
         $filter->method('getMatchingIds')->willReturn($result);
 
         return $filter;
