@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2021 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -14,7 +14,7 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2021 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -23,10 +23,13 @@ namespace MetaModels\Test\Helper;
 
 use Doctrine\DBAL\Connection;
 use MetaModels\Helper\TableManipulator;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
 /**
  * Class TableManipulatorTest
+ *
+ * @covers \MetaModels\Helper\TableManipulator
  */
 class TableManipulatorTest extends TestCase
 {
@@ -61,7 +64,7 @@ class TableManipulatorTest extends TestCase
     /**
      * Mock the database connection.
      *
-     * @return \PHPUnit_Framework_MockObject_MockObject|Connection
+     * @return MockObject|Connection
      */
     private function mockConnection()
     {
@@ -81,7 +84,7 @@ class TableManipulatorTest extends TestCase
     {
         $manipulator = $this->createTableManipulator();
 
-        $this->assertInstanceOf(TableManipulator::class, $manipulator);
+        self::assertInstanceOf(TableManipulator::class, $manipulator);
     }
 
     /**
@@ -98,11 +101,11 @@ class TableManipulatorTest extends TestCase
         $reservedWords = $property->getValue(TableManipulator::class);
 
         foreach ($reservedWords as $word) {
-            $this->assertTrue($manipulator->isReservedWord($word));
+            self::assertTrue($manipulator->isReservedWord($word));
         }
 
         foreach (['foo', 'bar'] as $word) {
-            $this->assertFalse($manipulator->isReservedWord($word));
+            self::assertFalse($manipulator->isReservedWord($word));
         }
     }
 }
