@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2021 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -14,7 +14,7 @@
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2021 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -47,8 +47,8 @@ class ToolBoxFileTest extends TestCase
             'meta'  => []
         ];
 
-        $this->assertEquals($emptyExpected, ToolboxFile::convertUuidsOrPathsToMetaModels(null));
-        $this->assertEquals($emptyExpected, ToolboxFile::convertUuidsOrPathsToMetaModels([]));
-        $this->assertEquals($emptyExpected, ToolboxFile::convertUuidsOrPathsToMetaModels([null]));
+        self::assertEquals($emptyExpected, ToolboxFile::convertUuidsOrPathsToMetaModels(null));
+        self::assertEquals($emptyExpected, ToolboxFile::convertUuidsOrPathsToMetaModels([]));
+        self::assertEquals($emptyExpected, ToolboxFile::convertUuidsOrPathsToMetaModels([null]));
     }
 }
